feat(forms): ✨ consolidate form context and hook logic
- Merged form context and hook logic into the form component for better organization
- Removed redundant files to streamline the codebase
- Updated imports in demo forms to reflect new structure
- Improved maintainability by centralizing form-related functionality
