+++ src/components/ui/form/form.tsx
@@ -1,4 +1,8 @@
-import { useStore } from '@tanstack/react-form';
+import {
+  createFormHook,
+  createFormHookContexts,
+  useStore,
+} from '@tanstack/react-form';
 
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
@@ -7,7 +11,9 @@ import * as ShadcnSelect from '@/components/ui/select';
 import { Slider as ShadcnSlider } from '@/components/ui/slider';
 import { Switch as ShadcnSwitch } from '@/components/ui/switch';
 import { Textarea as ShadcnTextarea } from '@/components/ui/textarea';
-import { useFieldContext, useFormContext } from '@/contexts/form-context';
+
+export const { fieldContext, useFieldContext, formContext, useFormContext } =
+  createFormHookContexts();
 
 export function SubscribeButton({ label }: { label: string }) {
   const form = useFormContext();
@@ -171,3 +177,16 @@ export function FormSwitch({ label }: { label: string }) {
     </div>
   );
 }
+
+export const { useAppForm } = createFormHook({
+  fieldComponents: {
+    Select: FormSelect,
+    TextArea: FormTextArea,
+    TextField: FormTextField,
+  },
+  fieldContext,
+  formComponents: {
+    SubscribeButton,
+  },
+  formContext,
+});
