+++ src/routes/_app/demos/forms/date-picker.tsx
@@ -9,7 +9,7 @@ import {
   CardTitle,
 } from '@/components/ui/card';
 import { DatePicker } from '@/components/ui/date-picker';
-import { useAppForm } from '@/hooks/use-app-form';
+import { useAppForm } from '@/components/ui/form';
 
 export const Route = createFileRoute('/_app/demos/forms/date-picker')({
   component: DatePickerForm,
