+++ src/routes/_app/demos/forms/date-range-picker.tsx
@@ -8,7 +8,7 @@ import {
   CardTitle,
 } from '@/components/ui/card';
 import { DateRangePicker } from '@/components/ui/date-range-picker';
-import { useAppForm } from '@/hooks/use-app-form';
+import { useAppForm } from '@/components/ui/form';
 
 export const Route = createFileRoute('/_app/demos/forms/date-range-picker')({
   component: DateRangePickerForm,
